+++ bin/binary_generator.dart
@@ -4,6 +4,7 @@ import "dart:io";
 
 import "package:args_helper/args_helper.dart";
 import "package:binary_generator/library_generator/library_generator.dart";
+import "package:binary_generator/types_generator/types_generator.dart";
 import "package:strings/strings.dart";
 import "package:path/path.dart" as pathos;
 import "package:yaml/yaml.dart" as yaml;
@@ -23,6 +24,23 @@ class Program {
     }
 
     var options = new LibraryGeneratorOptions(header: header, library: library, name: name);
+    var lines = generator.generate(options);
+    if (output == null) {
+      output = name.toLowerCase() + ".dart";
+    }
+
+    _writeToFile(output, lines.join("\n"));
+  }
+
+  void typesCommand(String filename, {String library, String name, String output}) {
+    var header = _readFromFile(filename);
+    name = _getClassName(filename, name);
+    var generator = new TypesGenerator(header);
+    if (library == null) {
+      library = name.toLowerCase();
+    }
+
+    var options = new TypesGeneratorOptions(header: header, library: library, name: name);
     var lines = generator.generate(options);
     if (output == null) {
       output = name.toLowerCase() + ".dart";
@@ -61,6 +79,19 @@ const _config = '''
 name: binary_generator
 description: Binary generator for develop code from the binary interop and binary types.
 commands:
+  types:
+    description: Generate the binary types code for typedef types 
+    options:       
+      name:        
+        help: Name of the generated Dart class (eg. --name MyTypes)                
+      library:        
+        help: Name of the generated Dart library (eg. --library libs.mytypes)
+      output:        
+        help: Output file name (eg. --output mytypes.dart)
+    rest:
+      allowMultiple: false
+      name: filename
+      required: true
   library:
     description: Generate the library code for functions and structures 
     options: