+++ lib/internal/type_helper/type_helper.dart
@@ -5,4 +5,4 @@ import "dart:convert";
 import "package:binary_declarations/binary_declarations.dart";
 import "package:binary_types/binary_types.dart";
 
-part "src/type_converter.dart";
+part "src/type_helper.dart";