+++ lib/types_generator/types_generator.dart
@@ -1,4 +1,4 @@
-library binary_generator.library_generator;
+library binary_generator.types_generator;
 
 import "package:binary_declarations/binary_declarations.dart";
 import "package:binary_generator/internal/generators/generators.dart";
@@ -7,7 +7,7 @@ import "package:binary_types/binary_types.dart";
 import "package:template_block/template_block.dart";
 
 part 'src/constructor_generator.dart';
-part 'src/class_library_generator.dart';
-part 'src/foreign_function_generator.dart';
-part 'src/library_generator.dart';
+part 'src/class_types_generator.dart';
+part 'src/getter_type_generator.dart';
+part 'src/types_generator.dart';
 part 'src/script_generator.dart';