+++ CHANGELOG.md
@@ -1,6 +1,19 @@
+## 0.0.5
+
+- Added new generator `TypesGenerator`. The generator allows to generate the `binary types` from the textual format supported by the `binary_declarations`. The declarations "typedef"  will be represented as the members of the generated class `binary types`   
+
+## 0.0.4
+
+- Made adaptations to the new version of package `binary_declarations`
+- Made adaptations to the new version of package `binary_types`
+
+## 0.0.3
+
+- Fixec bug in `ForeignFunctionGenerator`
+
 ## 0.0.2
 
-- Library generator now utilizes the features of `binary types` not only `binary declarations`. For the better type resolutions now generator tries to declare the its own copy of the `binary types` directly from the `binary declarations`  
+- Library generator now utilizes the features of `binary types` not only `binary declarations`. For the better type resolutions now generator tries to declare the its own copy of the `binary types` directly from the `binary declarations`
 
 ## 0.0.1